Factor out replay code into a library
I want to add a version of the benchmark which produces a Daml profile
and don’t want to duplicate all the code.

changelog_begin
changelog_end
